+++ DesignToolbox/DesignToolbox/Pages/Components/Switch/Switchitem/SwitchItemConfiguration.swift
@@ -133,42 +133,28 @@ struct SwitchItemConfiguration: View {
 
     var body: some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedMedium) {
-            Toggle("app_components_switch_selection_label", isOn: $model.selection)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-                .disabled(model.isError || model.isReadOnly)
+            VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
+                OUDSSwitchItem("app_components_switch_selection_label", isOn: $model.selection)
+                    .disabled(model.isError || model.isReadOnly)
 
-            Toggle("app_components_controlItem_icon_label", isOn: $model.icon)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+                OUDSSwitchItem("app_components_controlItem_icon_label", isOn: $model.icon)
 
-            Toggle("app_components_controlItem_flipIcon_label", isOn: $model.flipIcon)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-                .disabled(!model.icon)
+                OUDSSwitchItem("app_components_controlItem_flipIcon_label", isOn: $model.flipIcon)
+                    .disabled(!model.icon)
 
-            Toggle("app_components_controlItem_divider_label", isOn: $model.divider)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+                OUDSSwitchItem("app_components_controlItem_divider_label", isOn: $model.divider)
 
-            Toggle("app_components_controlItem_reversed_label", isOn: $model.isReversed)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+                OUDSSwitchItem("app_components_controlItem_reversed_label", isOn: $model.isReversed)
 
-            Toggle("app_common_enabled_label", isOn: $model.enabled)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-                .disabled(model.isError || model.isReadOnly)
+                OUDSSwitchItem("app_common_enabled_label", isOn: $model.enabled)
+                    .disabled(model.isError || model.isReadOnly)
 
-            Toggle("app_components_controlItem_readOnly_label", isOn: $model.isReadOnly)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-                .disabled(!model.enabled || model.isError)
+                OUDSSwitchItem("app_components_controlItem_readOnly_label", isOn: $model.isReadOnly)
+                    .disabled(!model.enabled || model.isError)
 
-            Toggle("app_components_common_error_label", isOn: $model.isError)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-                .disabled(!model.enabled || model.isReadOnly)
+                OUDSSwitchItem("app_components_common_error_label", isOn: $model.isError)
+                    .disabled(!model.enabled || model.isReadOnly)
+            }
 
             DesignToolboxEditContentDisclosure {
                 DesignToolboxTextField(text: $model.labelText)